\name{clustersumm}
\alias{clustersumm}
clustersumm.
\title{
Returns summary of information in each cluster
}
\description{
This function allows the user to return the information coded in each cluster, showing numerical values on these observations.
}
\usage{
clustersumm(df,k)
}
% - It needs to have installed the package cluster,Rtsne,ggplot2,tidyr.
\arguments{
  \item{df}{
%%     ~~This item will be the df formatted with the transformation function~~

}
  \item{k}{
%% ~~ this item will be the number of cluster selected by the user~~
  }
}
\details{
%%  ~~ It returns a list of each cluster, with a summary of their values. ~~
}
\value{
%%  ~It returns a list ~~
}
\references{
%% ~~ggplot2 : Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York~~
%% ~~"Finding Groups in Data": Cluster Analysis Extended Rousseeuw et al. [R package cluster version 2.1.4]. (2022, August 22). Comprehensive R Archive Network (CRAN). Retrieved from https://cran.r-project.org/web/packages/cluster/index.html~~
%% ~~Krijthe, J. (2022). T-Distributed Stochastic Neighbor Embedding using a Barnes-Hut Implementation [R package Rtsne version 0.16]. Comprehensive R Archive Network (CRAN). Retrieved from https://cran.r-project.org/web/packages/Rtsne/index.html~~
%% ~~ Wickham H, Girlich M (2022). tidyr: Tidy Messy Data. https://tidyr.tidyverse.org, https://github.com/tidyverse/tidyr. ~~

}
\author{
%%  ~~Jorge Miguel Alcantara Espinosa~~
}
\note{
%%  ~~Please report any bugs to jorge.alcantara.espinosa@hotmail.com~~
}
