\name{kcluster}
\alias{kcluster}
kcluster.
\title{
Returns graphical plot of recommended n clusters
}
\description{
This function allows the user to display in a graphical plot the n cluster and the silhouette widht related to them.
}
\usage{
kcluster(df)
}
%-It needs to have installed the package cluster,Rtsne,ggplot2.
\arguments{
  \item{df}{
%%     ~~This item will be the df formatted with the transformation function~~

}
}
\details{
%%  ~~ It returns a silhouette width analysis. ~~
}
\value{
%%  ~It returns a graphical plot ~~
}
\references{
%% ~~ggplot2 : Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York~~
%% ~~"Finding Groups in Data": Cluster Analysis Extended Rousseeuw et al. [R package cluster version 2.1.4]. (2022, August 22). Comprehensive R Archive Network (CRAN). Retrieved from https://cran.r-project.org/web/packages/cluster/index.html~~
%% ~~Krijthe, J. (2022). T-Distributed Stochastic Neighbor Embedding using a Barnes-Hut Implementation [R package Rtsne version 0.16]. Comprehensive R Archive Network (CRAN). Retrieved from https://cran.r-project.org/web/packages/Rtsne/index.html~~

}
\author{
%%  ~~Jorge Miguel Alcantara Espinosa~~
}
\note{
%%  ~~Please report any bugs to jorge.alcantara.espinosa@hotmail.com~~
}
