\name{predictors}
\alias{predictors}
%- Predictors.
\title{
%%  ~~Returns predictors of the model ~~
}
\description{
%%  ~~ This function allows the user to get given predictors of the relation between needing reevaluation and the given variables. ~~
}
\usage{
predictors(df,n)
}
\arguments{
  \item{df}{
%%     ~~This item will be the formatted df~~

}
  \item{n}{
%%     ~~This item will be the number of predictors we need~~

}
}
\details{
%%  ~~ This function will create a lineal model with the variables of the data.frame, then it will look which variables are more relevant to the need of reevaluation, it will sort them by importance and show them as a list~~
}
\value{
%%  ~It returns a list of predictors
}
\references{
}
\author{
%%  ~~Jorge Miguel Alcántara Espinosa~~
}
\note{
}
