\name{transformation}
\alias{transformation}
Transformation.
\title{
Transform the df to a coded data.frame
}
\description{
%%  ~~ This function allows the user to transform a given dataframe assigned to the variable df to a data.frame coded as needed by the other functions. ~~
}
\usage{
visualization(df)
}
\arguments{
  \item{df}{
%%     ~~This item will be the df that will be formatted~~

}
}
\details{
%%  ~~ This function will change the data.frame to a data.frame with 12 columns. These columns will be "center","act","prof",                        "gender","age","location","actw","acts","wday"","hday","census","nday" ~~
}
\value{
%%  ~It returns an updated df
}
\references{
}
\author{
%%  ~~Jorge Miguel Alcántara Espinosa~~
}
\note{
}
