\name{visualization}
\alias{visualization}
Visualization.
\title{
Visualize relation between variables
}
\description{
This function allows the user to display different graphical tools to be able to visualize relations between variables
}
\usage{
visualization(df,y)
}
%-It needs to have installed the package ggplot.
\arguments{
  \item{df}{
%%     ~~This item will be the df formatted with the transformation function~~

}
  \item{y}{
%% ~~ this item will be the plot selected by the user~~
  }
}
\details{
%%  ~~ There are 6 plots that the user can choose from ~~
}
\value{
%%  ~It returns a graphical plot
}
\references{
%% ~~ggplot2 : Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York~
}
\author{
%%  ~~Jorge Miguel Alcántara Espinosa~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

